Update Course.h
THis project works as a class registration of a student in college.  The student can add choose his or her class that they which  
to take and it is added to their cart.  If they choose to withdraw, that is updated.  The same is for a teacher. 
The student class interacts with the teacher and 
class lists to uddate those as well.